+++ client/src/components/recruit-map/KakaoMap.jsx
@@ -1,15 +1,13 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
-import { Map, MapMarker, useMap } from 'react-kakao-maps-sdk';
+import { Map } from 'react-kakao-maps-sdk';
 
 import { regionAtom, targetCafeAtom, scopeAtom, cafeInfoAtom } from '../../recoil/recruit-map';
-import { ResizeButtonContainer } from './ResizeButtonContainer';
-import { InfoWindow } from './InfoWindow';
 import { regionCoordinate } from '../../constants/regionCoordinate';
 import { ApiUrl } from '../../constants/ApiUrl';
+import Marker from './Marker';
+import { ResizeButtonContainer } from './ResizeButtonContainer';
 import * as api from '../../utils/api';
-import markerImg from '../../assets/images/icon/marker.png';
-import clickedMarkerImg from '../../assets/images/icon/marker-clicked.png';
 
 async function getCafeInfo(region) {
   try {
@@ -35,12 +33,9 @@ const getMapScreenScope = (map) => {
 
 export default function KakaoMap() {
   const mapRef = useRef();
-  const BASIC_SIZE = 40;
-  const OVER_SIZE = 42;
   const region = useRecoilValue(regionAtom);
-  const [targetCafe, setTargetCafe] = useRecoilState(targetCafeAtom);
-  const [isClicked, setIsClicked] = useState(true);
   const [cafeInfo, setCafeInfo] = useRecoilState(cafeInfoAtom);
+  const setTargetCafe = useSetRecoilState(targetCafeAtom);
   const setScope = useSetRecoilState(scopeAtom);
   const [level, setLevel] = useState(5);
 
@@ -57,44 +52,6 @@ export default function KakaoMap() {
   useEffect(() => {
     handleRegionChange();
   }, [region]);
-
-  const MarkerContainer = ({ cafeId, position, cafeName, recruitingNum }) => {
-    const map = useMap();
-    const [isOver, setIsOver] = useState(false);
-    let markerIcon = cafeId === targetCafe ? clickedMarkerImg : markerImg;
-    let markerSize = isOver ? OVER_SIZE : BASIC_SIZE;
-    const onMarkerClick = (marker, cafeId) => {
-      map.panTo(marker.getPosition());
-      setIsClicked(!isClicked);
-      setTimeout(() => {
-        if (isClicked) return setTargetCafe(cafeId);
-        return setTargetCafe(undefined);
-      });
-    };
-    return (
-      <MapMarker
-        position={position}
-        image={{
-          src: markerIcon,
-          size: {
-            width: markerSize,
-            height: markerSize,
-          },
-        }}
-        clickable={true}
-        onClick={(marker) => {
-          onMarkerClick(marker, cafeId);
-        }}
-        onMouseOver={() => setIsOver(true)}
-        onMouseOut={() =>
-          setTimeout(function () {
-            setIsOver(false);
-          })
-        }>
-        {isOver && <InfoWindow cafeName={cafeName} recruitingNum={recruitingNum} />}
-      </MapMarker>
-    );
-  };
 
   return (
     <div className='relative'>
@@ -115,7 +72,7 @@ export default function KakaoMap() {
           map={mapRef.current}></ResizeButtonContainer>
         {cafeInfo?.[region] &&
           cafeInfo[region].map((cafe) => (
-            <MarkerContainer
+            <Marker
               key={cafe.cafeId}
               cafeId={cafe.cafeId}
               position={{ lat: cafe.lat, lng: cafe.lng }}
